requirements.txt
EasyProcess==0.1.6
PyVirtualDisplay==0.1.5
SQLAlchemy==0.9.8
Scrapy==0.24.4
Twisted==14.0.2
argparse==1.2.1
cffi==0.8.6
cryptography==0.6
cssselect==0.9.1
lxml==3.4.0
psycopg2==2.5.4
pyOpenSSL==0.14
pycparser==2.10
python-dateutil==2.2
queuelib==1.2.2
selenium==2.44.0
six==1.8.0
w3lib==1.10.0
wsgiref==0.1.2
xlwt==0.7.5
zope.interface==4.1.1
service-identity==14.0.0
